rename LitOptions -> LitShaderOptions and split out LitMaterialOptionsBuilder
