+++ src/deprecated/deprecated.js
@@ -77,7 +77,7 @@ import { StandardMaterial } from '../scene/materials/standard-material.js';
 import { Batch } from '../scene/batching/batch.js';
 import { getDefaultMaterial } from '../scene/materials/default-material.js';
 import { StandardMaterialOptions } from '../scene/materials/standard-material-options.js';
-import { LitOptions } from '../scene/materials/lit-options.js';
+import { LitShaderOptions } from '../scene/shader-lib/programs/lit-shader-options.js';
 import { Layer } from '../scene/layer.js';
 
 import { Animation, Key, Node } from '../scene/animation/animation.js';
@@ -117,6 +117,7 @@ import {
 import { RigidBodyComponent } from '../framework/components/rigid-body/component.js';
 import { RigidBodyComponentSystem } from '../framework/components/rigid-body/system.js';
 import { basisInitialize } from '../framework/handlers/basis.js';
+import { LitShader } from '../scene/shader-lib/programs/lit-shader.js';
 
 // CORE
 
@@ -679,6 +680,7 @@ GraphicsDevice.prototype.getCullMode = function () {
 // SCENE
 
 export const PhongMaterial = StandardMaterial;
+export const LitOptions = LitShaderOptions;
 
 export const scene = {
     partitionSkin: partitionSkin,
@@ -1018,11 +1020,11 @@ function _defineOption(name, newName) {
     if (name !== 'chunks' && name !== '_pass' && name !== '_isForwardPass') {
         Object.defineProperty(StandardMaterialOptions.prototype, name, {
             get: function () {
-                Debug.deprecated(`Getting pc.Options#${name} has been deprecated as the property has been moved to pc.Options.LitOptions#${newName || name}.`);
+                Debug.deprecated(`Getting pc.Options#${name} has been deprecated as the property has been moved to pc.Options.LitShaderOptions#${newName || name}.`);
                 return this.litOptions[newName || name];
             },
             set: function (value) {
-                Debug.deprecated(`Setting pc.Options#${name} has been deprecated as the property has been moved to pc.Options.LitOptions#${newName || name}.`);
+                Debug.deprecated(`Setting pc.Options#${name} has been deprecated as the property has been moved to pc.Options.LitShaderOptions#${newName || name}.`);
                 this.litOptions[newName || name] = value;
             }
         });
@@ -1030,7 +1032,7 @@ function _defineOption(name, newName) {
 }
 _defineOption('refraction', 'useRefraction');
 
-const tempOptions = new LitOptions();
+const tempOptions = new LitShaderOptions();
 const litOptionProperties = Object.getOwnPropertyNames(tempOptions);
 for (const litOption in litOptionProperties) {
     _defineOption(litOptionProperties[litOption]);
