+++ src/scene/shader-lib/programs/lit-shader.js
@@ -55,7 +55,7 @@ class LitShader {
         /**
          * @param {import('../../../platform/graphics/graphics-device.js').GraphicsDevice} device - The
          * graphics device.
-         * @param {import('../../../scene/materials/lit-options.js').LitOptions} options - The
+         * @param {import('./lit-shader-options.js').LitShaderOptions} options - The
          * lit options.
          * @ignore
          */
